Replace Storyboard+SwizzlingSpec with Storyboard+SwizzlingTests.
